去除循环中的隐形转换。

+++ 这是我见过最无聊的游戏/main.cpp
@@ -285,7 +285,7 @@ void DrawCircle (HWND hWnd)
 	m_pRenderTarget->Clear (D2D1::ColorF (D2D1::ColorF::Black));
 
 	// int count = circles.size (); // 之前为了减少每次执行size浪费的时间，所以在循环前先获取大小
-	for (int index = 0; index < circles.size (); ++index)
+	for (size_t index = 0; index < circles.size (); ++index)
 	{
 		Circle* c = circles[index];
 		if (nullptr == c)
